refactor(keystone): DOMA-13081 extract DEFAULT_PROCESSORS to separate module and simplify resolveExternalContentValue
- Move DEFAULT_PROCESSORS from Implementation.js to new defaultProcessors.js module
- Remove serialize/deserialize from ExternalContentImplementation constructor params
- Use formatProcessors[format].serialize instead of this.serialize in resolveInput
- Merge custom processors with DEFAULT_PROCESSORS in resolveExternalContentValue
- Remove deserialize, fieldPath, and item params from resolveExternalContentValue()
