+++ packages/keystone/fieldsUtils/ExternalContent/index.js
@@ -1,5 +1,6 @@
 const { FILE_META_TYPE } = require('./constants')
 const { createExternalDataField } = require('./createExternalDataField')
+const { DEFAULT_PROCESSORS } = require('./defaultProcessors')
 const { FileContentLoader } = require('./FileContentLoader')
 const { getOrCreateLoader } = require('./getOrCreateLoader')
 const { isFileMeta } = require('./isFileMeta')
@@ -10,6 +11,7 @@ module.exports = {
     FILE_META_TYPE,
     FileContentLoader,
     createExternalDataField,
+    DEFAULT_PROCESSORS,
     getOrCreateLoader,
     isFileMeta,
     resolveExternalContentValue,
